% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{L5_WINDOW_MINUTES}
\alias{L5_WINDOW_MINUTES}
\title{Window length (in minutes) used for the L5 statistic: average activity
over the least active 5-hour span of the average day.}
\format{
An object of class \code{integer} of length 1.
}
\usage{
L5_WINDOW_MINUTES
}
\description{
Window length (in minutes) used for the L5 statistic: average activity
over the least active 5-hour span of the average day.
}
\keyword{internal}
