% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{M10_WINDOW_MINUTES}
\alias{M10_WINDOW_MINUTES}
\title{Window length (in minutes) used for the M10 statistic: average activity
over the most active 10-hour span of the average day.}
\format{
An object of class \code{integer} of length 1.
}
\usage{
M10_WINDOW_MINUTES
}
\description{
Window length (in minutes) used for the M10 statistic: average activity
over the most active 10-hour span of the average day.
}
\keyword{internal}
