% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{MAX_AUTOCORR_LAG}
\alias{MAX_AUTOCORR_LAG}
\title{Autocorrelation lag (in minutes) beyond which "recovery" (autocorrelation
dropping below the 0.2 threshold) is capped, if it never actually drops
below threshold within the computed lags.}
\format{
An object of class \code{integer} of length 1.
}
\usage{
MAX_AUTOCORR_LAG
}
\description{
Autocorrelation lag (in minutes) beyond which "recovery" (autocorrelation
dropping below the 0.2 threshold) is capped, if it never actually drops
below threshold within the computed lags.
}
\keyword{internal}
