% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{MINUTES_PER_DAY}
\alias{MINUTES_PER_DAY}
\title{Minutes in a day, for a 1-minute-epoch actigraphy recording.}
\format{
An object of class \code{integer} of length 1.
}
\usage{
MINUTES_PER_DAY
}
\description{
Minutes in a day, for a 1-minute-epoch actigraphy recording.
}
\keyword{internal}
