% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/actman.r
\docType{data}
\name{ACTdata.1}
\alias{ACTdata.1}
\title{Data included in the package}
\description{
Data included in the package
}
\references{
\url{exampledata.com}
}
\author{
Yoram Kunkels \email{y.k.kunkels@umcg.nl}
}
\keyword{data}
