% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/actman.r
\name{actman}
\alias{actman}
\title{Loop script - Actiwatch 2 Respironics}
\usage{
actman(workdir)
}
\arguments{
\item{workdir}{The working directory of the script.}
}
\value{
nothing
}
\description{
This script loops and does some stuff.
}
\examples{
\dontrun{
actman("C:/Users/Y. K. Kunkels/Desktop/loopscript/datatest2")
}
}

