% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/actman.r
\name{ACTman}
\alias{ACTman}
\title{ACTman - Actigraphy Manager}
\usage{

  ACTman(workdir = "C:/Bibliotheek/Studie/PhD/Publishing/ACTman/R-part/mydata2",
  sleepdatadir = "C:/Bibliotheek/Studie/PhD/Publishing/ACTman/R-part/Actogram & Sleep analysis",
  myACTdevice = "Actiwatch2", iwantsleepanalysis = FALSE,
  plotactogram = FALSE, selectperiod = FALSE, startperiod, daysperiod,
  movingwindow = FALSE, lengthcheck = FALSE, na_omit = TRUE)
}
\arguments{
\item{workdir}{The working directory of the script.}

\item{sleepdatadir}{An optional vector specifying the directory for actogram and sleep analysis data.}

\item{myACTdevice}{Name of the input device used. Should be either 'Actiwatch2' or 'MW8'.}

\item{iwantsleepanalysis}{Boolean value indicating whether sleep analysis should be performed.}

\item{plotactogram}{Boolean value indicating whether an actogram has to be plotted.}

\item{selectperiod}{Boolean value indicating whether a specific period has to be selected.}

\item{startperiod}{An optional vector specifying single or multiple period starts. Should be in the format "2016-10-03 00:00:00".}

\item{daysperiod}{An optional vector specifying the length in days of the period.}

\item{movingwindow}{Boolean value indicating whether a moving window should be utilised.}

\item{lengthcheck}{Boolean value indicating whether data recorded after 14 days should be included.}

\item{na_omit}{Boolean value indicating whether NA's should be omitted.}
}
\value{
nothing
}
\description{
ACTman manages actigraphy data whilst offering pre-processing and analyses options.
This initial version supports the 'Actiwatch 2' and 'MotionWatch 8' actigraphy devices,
whilst allowing for both sleep and circadian rhythm analyses.
}
\examples{
\dontrun{
View(ACTman::ACTman(workdir = "C:/Bibliotheek/Studie/PhD/Publishing/ACTman/R-part/mydata2",
                    myACTdevice = "Actiwatch2",
                    iwantsleepanalysis = FALSE,
                    plotactogram = FALSE,
                    selectperiod = FALSE,
                    startperiod = "2016-10-03 00:00:00",
                    daysperiod = 14, movingwindow = FALSE, lengthcheck = TRUE))
}
}
