% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nparcalc.R
\name{nparcalc}
\alias{nparcalc}
\title{Nparcalc}
\usage{
nparcalc(lastwhole24h.pos, newdir, myACTdevice, ACTdata_file, plotactogram)
}
\arguments{
\item{lastwhole24h.pos}{Position in the data of the last whole 24 hour block.}

\item{newdir}{Working directory.}

\item{myACTdevice}{Name of the input device used. Should be either Actiwatch2 or MW8.}

\item{ACTdata_file}{List of the actigraphy data files.}

\item{plotactogram}{Boolean value indicating whether or not to plot an actogram.}
}
\value{
A list with the result values IS, IV, RA, L5, L5_starttime, M10, and M10_starttime.
}
\description{
Calculate non-parametrical circadian rhythm variables, such as: IS, IV, L5, and M10.
}
