% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nparcalc.R
\name{nparcalc}
\alias{nparcalc}
\title{Nparcalc}
\usage{
nparcalc(myACTdevice, movingwindow, CRV.data)
}
\arguments{
\item{myACTdevice}{Name of the input device used. Should be either Actiwatch2 or MW8.}

\item{movingwindow}{A boolean indicating whether moving window is used.}

\item{CRV.data}{CRV data}
}
\value{
A list with the result values IS, IV, RA, L5, L5_starttime, M10, and M10_starttime.
}
\description{
Calculate non-parametrical circadian rhythm variables, such as: IS, IV, L5, and M10.
}
