% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_actogram.R
\name{plot_actogram}
\alias{plot_actogram}
\title{plot_actogram}
\usage{
plot_actogram(CRV_data, workdir)
}
\arguments{
\item{CRV_data}{is the CRV.data object.}

\item{workdir}{the working directory as supplied to ACTman.}
}
\description{
Plots the actograms.
}
