% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_actogram.R
\name{plot_actogram}
\alias{plot_actogram}
\title{plot_actogram}
\usage{
plot_actogram(workdir, ACTdata.1.sub, i, plotactogram)
}
\arguments{
\item{workdir}{the working directory as supplied to ACTman.}

\item{ACTdata.1.sub}{the managed data set}
}
\description{
Function to plot 48 hour Actograms.
}
